Refactor metric popups
